requirements.txt
click==6.6
Django==1.8.14
django-rq==0.8.0
djangorestframework==3.4.7
ecdsa==0.13
Fabric==1.12.0
httpie==0.9.6
paramiko==1.17.2
pbr==1.10.0
pycrypto==2.6.1
Pygments==2.1.3
redis==2.10.5
requests==2.11.1
rq==0.6.0
six==1.10.0
stevedore==1.17.1
virtualenv==15.0.3
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
yolk==0.4.3
